requirements.txt
streamlit>=1.28.0
anthropic>=0.18.0
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
pdfplumber>=0.11.0
python-docx>=1.0.0
openpyxl>=3.1.0
playwright>=1.40.0
